refactor(reservation): WaitingRankReservation 파일 위치 변경 및 관련 경로 수정
WaitingRankReservation 클래스의 패키지를 dto에서 domain으로 이동하고, 관련된 클래스 및 쿼리 경로를 수정함.
