+++ src/main/java/roomescape/reservation/repository/ReservationJpaRepository.java
@@ -7,7 +7,7 @@ import roomescape.member.domain.Member;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationStatus;
-import roomescape.reservation.dto.WaitingRankReservation;
+import roomescape.reservation.domain.WaitingRankReservation;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 
@@ -57,7 +57,7 @@ public interface ReservationJpaRepository extends JpaRepository<Reservation, Lon
     );
 
     @Query(value = """
-    SELECT  new roomescape.reservation.WaitingRankReservation(r, 
+    SELECT  new roomescape.reservation.domain.WaitingRankReservation(r, 
         (
             SELECT COUNT(r2)
             FROM Reservation r2
